refactor: ReportBox를 Box와 Layout으로 컴포넌트 분리
