+++ src/pages/reports/Report/components/ReportLayout/index.ts
@@ -1,0 +1,3 @@
+import ReportLayout from "./ReportLayout.tsx";
+
+export default ReportLayout;
